+++ cli/src/command/mod.rs
@@ -15,7 +15,7 @@
 
 //! Define the commands supported by the CLI.
 
-use crate::{CommandError, CommandT, NetworkOptions, TransactionError, TxOptions};
+use crate::{CommandError, CommandT, NetworkOptions, TxOptions};
 use radicle_registry_client::*;
 
 use sp_core::crypto::Ss58Codec;
@@ -26,22 +26,6 @@ pub mod org;
 pub mod other;
 pub mod project;
 pub mod user;
-
-/// Check that a transaction has been applied succesfully.
-///
-/// Return the `Ok` value of the transaction result or, if the transaction failed
-/// (i.e., `tx_applied.result` is `Err`), return a [TransactionError].
-fn transaction_applied_ok<Message_, T>(
-    tx_applied: &TransactionApplied<Message_>,
-) -> Result<T, TransactionError>
-where
-    Message_: Message<Result = Result<T, DispatchError>>,
-    T: Copy + Send + 'static,
-{
-    tx_applied
-        .result
-        .map_err(|dispatch_error| dispatch_error.into())
-}
 
 fn parse_account_id(data: &str) -> Result<AccountId, String> {
     Ss58Codec::from_ss58check(data).map_err(|err| format!("{:?}", err))