+++ cli/src/command/org.rs
@@ -121,8 +121,7 @@ impl CommandT for Register {
             .await?;
         announce_tx("Registering org...");
 
-        let org_registered = register_org_fut.await?;
-        transaction_applied_ok(&org_registered)?;
+        register_org_fut.await?.result?;
         println!("✓ Org {} is now registered.", self.org_id);
         Ok(())
     }
@@ -156,8 +155,7 @@ impl CommandT for Unregister {
             .await?;
         announce_tx("Unregistering org...");
 
-        let org_unregistered = register_org_fut.await?;
-        transaction_applied_ok(&org_unregistered)?;
+        register_org_fut.await?.result?;
         println!("✓ Org {} is now unregistered.", self.org_id);
         Ok(())
     }
@@ -201,7 +199,7 @@ impl CommandT for Transfer {
         announce_tx("Transferring funds...");
 
         let transfered = transfer_fut.await?;
-        transaction_applied_ok(&transfered)?;
+        transfered.result?;
         println!(
             "✓ Transferred {} μRAD from Org {} to Account {} in block {}",
             self.amount, self.org_id, self.recipient, transfered.block,